+++ src/Classes/Remote/fsi6.h
@@ -9,17 +9,15 @@ enum Input{Roll = 0, Pitch = 1, Velocity = 2, Yaw = 3};
 class fsi6
 {
     private:
-        static const int validChannels = 5;
-        static const int channelAmount = 6;
-        static const unsigned long blankTime = 2100;
+        static const int8_t validChannels = 5;
+        static const int8_t channelAmount = 6;
+        static const uint32_t blankTime = 2100;
     public:
         float pitch_offset, roll_offset, yaw_offset;
         fsi6();
         static void isr();
         void calibrateRemote();
         static float InputValue[validChannels];
-
-        //void fillArray();
 };
